1,4 interactionsa and dihedrals
In the previous vrsion of the script, I could observe the volume expansion of simulation box by 25%. This problem occurred due to missing 1,4-interactions. The scripts creates them now. Even though all the dihedrals in the InterfaceFF are zero at the moment, they are also created by the script. According to my tests, these 2 additions to the code remove the problem of the simulation box expansion.